+++ src/main/java/net/mohron/skyclaims/command/team/CommandInvite.java
@@ -19,31 +19,48 @@
 package net.mohron.skyclaims.command.team;
 
 import net.mohron.skyclaims.command.CommandBase;
+import net.mohron.skyclaims.command.CommandIsland;
 import net.mohron.skyclaims.permissions.Permissions;
+import net.mohron.skyclaims.team.Invite;
+import net.mohron.skyclaims.team.InviteService;
+import net.mohron.skyclaims.team.PrivilegeType;
+import net.mohron.skyclaims.world.Island;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
-import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
 import org.spongepowered.api.command.args.GenericArguments;
 import org.spongepowered.api.command.spec.CommandSpec;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.entity.living.player.User;
 import org.spongepowered.api.text.Text;
+import org.spongepowered.api.text.format.TextColors;
+import org.spongepowered.api.text.format.TextStyles;
 
-public class CommandInvite extends CommandBase {
+import javax.annotation.Nonnull;
+
+public class CommandInvite extends CommandBase.IslandCommand {
 
     public static final String HELP_TEXT = "used to invite players to your island or list your pending invites.";
+    private static final Text LIST = Text.of("list");
     private static final Text USER = Text.of("user");
+    private static final Text PRIVILEGE = Text.of("privilege");
 
     public static CommandSpec commandSpec = CommandSpec.builder()
         .permission(Permissions.COMMAND_INVITE)
-        .arguments(GenericArguments.optional(GenericArguments.user(USER)))
+        .arguments(
+            GenericArguments.optional(GenericArguments.seq(
+                GenericArguments.user(USER),
+                GenericArguments.optional(PrivilegeType.getCommandArgument(PRIVILEGE))
+            )),
+            GenericArguments.optional(GenericArguments.literal(LIST, "list"))
+        )
         .description(Text.of(HELP_TEXT))
         .executor(new CommandInvite())
         .build();
 
     public static void register() {
         try {
+            CommandIsland.addSubCommand(commandSpec, "invite");
             PLUGIN.getGame().getCommandManager().register(PLUGIN, commandSpec);
             PLUGIN.getLogger().debug("Registered command: CommandInvite");
         } catch (UnsupportedOperationException e) {
@@ -52,22 +69,41 @@ public class CommandInvite extends CommandBase {
         }
     }
 
-    @Override
-    public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
-        if (!(src instanceof Player)) {
-            throw new CommandException(Text.of("You must be a player to run this command!"));
-        }
-        Player player = (Player) src;
-        User user = (User) args.getOne(USER).orElse(null);
+    @Override public CommandResult execute(@Nonnull Player player, @Nonnull Island island, @Nonnull CommandContext args) throws CommandException {
+        User user = args.<User>getOne(USER).orElse(null);
+        PrivilegeType type = args.<PrivilegeType>getOne(PRIVILEGE).orElse(PrivilegeType.MEMBER);
 
-        if (user == null) {
-            //TODO List invites
+        if (type == PrivilegeType.NONE) {
+            throw new CommandException(
+                Text.of(TextStyles.ITALIC, "What kind of invite is ", TextStyles.RESET, type.toText(), TextStyles.ITALIC, "?"));
+        }
+
+        if (user == null || args.hasAny(LIST)) {
+            PLUGIN.getInviteService().listIncomingInvites().accept(player);
+        } else if (player.equals(user)) {
+            throw new CommandException(Text.of(TextColors.RED, "You cannot invite yourself!"));
+        } else if (island.getPrivilegeType(user) != PrivilegeType.NONE) {
+            throw new CommandException(Text.of(
+                island.getPrivilegeType(user).format(user.getName()), TextColors.RED, " is already a ", island.getPrivilegeType(user).toText(),
+                TextColors.RED, " on ", island.getName(), TextColors.RED, "!"
+            ));
+        } else if (!island.isOwner(player) || island.getPrivilegeType(player).ordinal() >= type.ordinal()) {
+            throw new CommandException(Text.of(
+                TextColors.RED, "You do not have permission to send ", type == PrivilegeType.OWNER ? "an " : "a ",
+                type.toText(), TextColors.RED, " invite for ", island.getName(), TextColors.RED, "!"
+            ));
         } else {
-            //TODO Send invite to online player or save for login
+            Invite.builder()
+                .island(island)
+                .sender(player)
+                .receiver(user)
+                .privilegeType(type)
+                .build()
+                .send();
+            player.sendMessage(Text.of(TextColors.GREEN, "Island invite sent to ", type.format(user.getName()), TextColors.GREEN, "."));
         }
 
-        throw new CommandException(Text.of("Command not yet implemented"));
-
-        //return CommandResult.success();
+        return CommandResult.success();
     }
+
 }