+++ src/main/java/net/mohron/skyclaims/command/user/CommandList.java
@@ -18,7 +18,6 @@
 
 package net.mohron.skyclaims.command.user;
 
-import com.google.common.collect.Lists;
 import net.mohron.skyclaims.SkyClaims;
 import net.mohron.skyclaims.command.CommandBase;
 import net.mohron.skyclaims.command.CommandIsland;
@@ -43,6 +42,7 @@ import org.spongepowered.api.text.format.TextStyles;
 import java.util.Comparator;
 import java.util.List;
 import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 public class CommandList extends CommandBase {
 
@@ -55,7 +55,7 @@ public class CommandList extends CommandBase {
         .description(Text.of(HELP_TEXT))
         .arguments(GenericArguments.firstParsing(
             GenericArguments.optional(GenericArguments.user(USER)),
-            GenericArguments.optional(Argument.sort(SORT))
+            GenericArguments.optional(GenericArguments.requiringPermission(Argument.sort(SORT), Permissions.COMMAND_LIST_SORT))
         ))
         .executor(new CommandList())
         .build();
@@ -71,102 +71,112 @@ public class CommandList extends CommandBase {
         }
     }
 
-    private static Consumer<CommandSource> createReturnConsumer(CommandSource src) {
-        return consumer -> {
-            Text returnCommand = Text.builder().append(Text.of(
-                Text.NEW_LINE, TextColors.WHITE, "[", TextColors.AQUA, "Return to Island List", TextColors.WHITE, "]", Text.NEW_LINE))
-                .onClick(TextActions.executeCallback(
-                    CommandUtil.createCommandConsumer(src, "islandlist", "", null))
-                ).build();
-            src.sendMessage(returnCommand);
-        };
-    }
-
     public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
         if (SkyClaims.islands.isEmpty()) {
             src.sendMessage(Text.of(TextColors.RED, "There are currently no islands!"));
             return CommandResult.empty();
         }
-        List<Text> listText = Lists.newArrayList();
         Player player = (src instanceof Player) ? (Player) src : null;
         User user = args.<User>getOne(USER).orElse(null);
         Comparator<Island> sortType = args.<Comparator<Island>>getOne(SORT).orElse(Comparator.comparing(Island::getSortableName));
 
-        boolean spawnOthers = src.hasPermission(Permissions.COMMAND_SPAWN_OTHERS);
+        boolean showUnlocked = src.hasPermission(Permissions.COMMAND_LIST_UNLOCKED);
+        boolean showAll = src.hasPermission(Permissions.COMMAND_LIST_ALL);
 
-        SkyClaims.islands.values().stream()
+        List<Text> listText = SkyClaims.islands.values().stream()
             .filter(i -> user == null || i.isMember(user))
+            .filter(i -> player == null || i.isMember(player) || !i.isLocked() && showUnlocked || showAll)
             .sorted(sortType)
-            .forEach(island -> listText.add(Text.of(
-                getLocked(island),
+            .map(island -> Text.of(
+                getAccess(island, src),
                 island.getName().toBuilder()
                     .onHover(TextActions.showText(Text.of("Click here to view island info")))
                     .onClick(TextActions.executeCallback(
-                        CommandUtil
-                            .createCommandConsumer(src, "islandinfo", island.getUniqueId().toString(), createReturnConsumer(src)))
-                    ),
-                (!island.isLocked() || ((player == null || island.isMember(player)) || spawnOthers))
-                    ? getClickableCoords(src, island) : getCoords(island)
-            )));
+                        CommandUtil.createCommandConsumer(src, "islandinfo", island.getUniqueId().toString(), createReturnConsumer())
+                    )),
+                getCoords(island, src)
+            ))
+            .collect(Collectors.toList());
 
         if (listText.isEmpty()) {
-            listText.add(Text.of(TextColors.RED, "There are no islands to display!"));
-        }
-
-        if (!(src instanceof Player)) {
-            listText.forEach(src::sendMessage);
-        } else {
+            src.sendMessage(Text.of(TextColors.RED, "There are no islands to display!"));
+        } else if (src instanceof Player) {
             PaginationList.builder()
                 .title(Text.of(TextColors.AQUA, "Island List"))
                 .padding(Text.of(TextColors.AQUA, TextStyles.STRIKETHROUGH, "-"))
                 .contents(listText)
                 .sendTo(src);
+        } else {
+            listText.forEach(src::sendMessage);
         }
 
         return CommandResult.success();
     }
 
-    private Text getLocked(Island island) {
-        return Text.of(TextColors.WHITE, " [",
-            Text.builder(island.isLocked() ? "L" : "U")
-                .color(island.isLocked() ? TextColors.RED : TextColors.GREEN)
-                .onHover(TextActions.showText(island.isLocked()
-                    ? Text.of(TextColors.RED, "LOCKED")
-                    : Text.of(TextColors.GREEN, "UNLOCKED")
-                ))
-                .onClick(TextActions.executeCallback(toggleLock(island))),
-            TextColors.WHITE, "] ");
+    private static Consumer<CommandSource> createReturnConsumer() {
+        return src -> src.sendMessage(Text.of(
+            TextActions.executeCallback(CommandUtil.createCommandConsumer(src, "islandlist", "", null)),
+            Text.NEW_LINE, TextColors.WHITE, "[",
+            TextColors.AQUA, "Return to Island List",
+            TextColors.WHITE, "]", Text.NEW_LINE
+        ));
+    }
+
+    private Text getAccess(Island island, CommandSource src) {
+        Text access;
+        Text hover = Text.EMPTY;
+        if (src instanceof Player && island.getOwnerUniqueId().equals(((Player) src).getUniqueId())) {
+            access = Text.of(TextColors.BLUE, "O");
+            hover = Text.of(TextColors.BLUE, "You own this Island", Text.NEW_LINE);
+        } else if (src instanceof Player && island.isManager((Player) src)) {
+            access = Text.of(TextColors.GOLD, "M");
+            hover = Text.of(TextColors.GOLD, "You are a Manager on this Island", Text.NEW_LINE);
+        } else if (src instanceof Player && island.isMember((Player) src)) {
+            access = Text.of(TextColors.YELLOW, "T");
+            hover = Text.of(TextColors.YELLOW, "You are Trusted on this Island", Text.NEW_LINE);
+        } else {
+            access = island.isLocked() ? Text.of(TextColors.RED, "L") : Text.of(TextColors.GREEN, "U");
+        }
+
+        hover = hover.concat(Text.of(island.getName(), TextColors.WHITE, " is ",
+            island.isLocked() ? Text.of(TextColors.RED, "Locked") : Text.of(TextColors.GREEN, "Unlocked")
+        ));
+
+        return Text.of(
+            TextColors.WHITE, " [",
+            access.toBuilder().onHover(TextActions.showText(hover)).onClick(TextActions.executeCallback(toggleLock(island))),
+            TextColors.WHITE, "] "
+        );
     }
 
     private Consumer<CommandSource> toggleLock(Island island) {
         return src -> {
             if (src instanceof Player
-                && ((Player) src).getUniqueId().equals(island.getOwnerUniqueId())
+                && island.isManager((Player) src)
                 && src.hasPermission(Permissions.COMMAND_LOCK)
                 || src.hasPermission(Permissions.COMMAND_LOCK_OTHERS)) {
                 island.setLocked(!island.isLocked());
-                src.sendMessage(Text.of(island.getName(), TextColors.GREEN, " is now ",
-                    Text.builder((island.isLocked()) ? "LOCKED" : "UNLOCKED")
+                src.sendMessage(Text.of(island.getName(), TextColors.GRAY, " is now ",
+                    Text.builder((island.isLocked()) ? "Locked" : "Unlocked")
                         .color((island.isLocked()) ? TextColors.RED : TextColors.GREEN)
                         .onClick(TextActions.executeCallback(toggleLock(island))),
-                    TextColors.GREEN, "!"
+                    TextColors.GRAY, "!"
                 ));
             }
         };
     }
 
-    private Text getCoords(Island island) {
-        return Text.of(TextColors.GRAY, " (",
+    private Text getCoords(Island island, CommandSource src) {
+        Text coords = Text.of(TextColors.GRAY, " (",
             TextColors.LIGHT_PURPLE, island.getRegion().getX(),
             TextColors.GRAY, ", ",
             TextColors.LIGHT_PURPLE, island.getRegion().getZ(),
             TextColors.GRAY, ")");
-    }
-
-    private Text getClickableCoords(CommandSource src, Island island) {
-        return getCoords(island).toBuilder()
+        return !island.isLocked() || src instanceof Player && island.isMember((Player) src) || src.hasPermission(Permissions.COMMAND_SPAWN_OTHERS)
+            ? coords.toBuilder()
             .onHover(TextActions.showText(Text.of("Click here to teleport to this island.")))
             .onClick(TextActions.executeCallback(CommandUtil.createTeleportConsumer(src, island.getSpawn().getLocation())))
-            .build();
+            .build()
+            : coords;
     }
 }