+++ src/main/java/net/mohron/skyclaims/config/type/MiscConfig.java
@@ -18,9 +18,11 @@
 
 package net.mohron.skyclaims.config.type;
 
+import net.mohron.skyclaims.SkyClaims;
 import ninja.leaping.configurate.objectmapping.Setting;
 import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -29,13 +31,28 @@ public class MiscConfig {
 
     @Setting(value = "Island-on-Join", comment = "Automatically create an island for a player on join.")
     private boolean islandOnJoin = false;
+    @Setting(value = "List-Schematics", comment = "Whether players with access to multiple schematics see a list when not specifying a schematic.")
+    private boolean listSchematics = true;
+    @Setting(value = "Teleport-on-Creation", comment = "Automatically teleport the owner to their island on creation.")
+    private boolean teleportOnCreate = true;
     @Setting(value = "Create-Commands", comment = "Commands to run on island creation and reset. Use @p in place of the player's name.")
     private List<String> createCommands = new ArrayList<>();
     @Setting(value = "Reset-Commands", comment = "Commands to run on island resets only. Use @p in place of the player's name.")
     private List<String> resetCommands = new ArrayList<>();
+    @Setting(value = "Date-Format", comment = "The date format used throughout the plugin.\n" +
+        "http://docs.oracle.com/javase/6/docs/api/java/text/SimpleDateFormat.html")
+    private String dateFormat = "MMMM d, yyyy h:mm a";
 
     public boolean createIslandOnJoin() {
         return islandOnJoin;
+    }
+
+    public boolean isListSchematics() {
+        return listSchematics;
+    }
+
+    public boolean isTeleportOnCreate() {
+        return teleportOnCreate;
     }
 
     public List<String> getCreateCommands() {
@@ -44,5 +61,14 @@ public class MiscConfig {
 
     public List<String> getResetCommands() {
         return resetCommands;
+    }
+
+    public SimpleDateFormat getDateFormat() {
+        try {
+            return new SimpleDateFormat(dateFormat);
+        } catch (IllegalArgumentException e) {
+            SkyClaims.getInstance().getLogger().info("Invalid Date Format: {}", dateFormat);
+            return new SimpleDateFormat("MMMM d, yyyy h:mm a");
+        }
     }
 }