+++ CHANGELOG.md
@@ -1,12 +1,28 @@
 # Change Log
 
-## Unreleased Features, Changes & Bugfixes
-**REQUIRED: SF build 2524+ & GP build 4.2.0.334+ OPTIONAL: Nucleus version 1.0.1+**
-- None
+## Upcoming/Unreleased Features, Changes & Bugfixes
+**REQUIRED: SF build 2558+ & GP build 4.0.0.415+ OPTIONAL: Nucleus version 1.1.3+**
+- Added Void World Generation capabilities
+- Added Nether portal fix
 
 # Beta 25 SNAPSHOT - S7.0
-**REQUIRED: SF build 2524+ & GP build 4.2.0.401+ OPTIONAL: Nucleus version 1.1.0+**
-- `/is create` & `/is reset` now support generating a clickable list of available schematics
+**REQUIRED: SF build 2624+ & GP build 4.2.0.418+ OPTIONAL: Nucleus version 1.2.0+**
+- Added island invite & rank system for working together with other players
+    - Added `/is invite` - `skyclaims.command.invite`
+    - Added `/is kick` - `skyclaims.command.kick`
+    - Added` /is leave` - `skyclaims.command.leave`
+    - Added` /is promote` - `skyclaims.command.promote`
+    - Added` /is demote` - `skyclaims.command.demote`
+    - Use of GriefPrevention's `/trust` & `/permissiontrust` on an island is handled by `/is invite`
+- Enhanced `/is list` and reworked permissions: 
+    - **[O]** - owner, **[M]** - manager, or **[T]** - trusted (member) now appear in place of **[L]** or **[T]** when applicable
+    - `skyclaims.command.list.base` - allows a player to use the list command _(shows a player's islands ie. **owner**, **manager**, or **trusted**)_
+    - `skyclaims.command.list.unlocked` - allows a player to list all unlocked islands
+    - `skyclaims.command.list.sort` - allows a player to use the list sorting features
+    - `skyclaims.admin.list` - allows a player to list all islands
+- Combined `/is info` owner & member sections & color coded entries
+- Added _Misc/Teleport-on-Creation_ config to disable automatic teleportation after island creation
+- `/is create` & `/is reset` now support generating a clickable list from available schematics
 - Added Void Island Control Schematics to prepackaged schematics
 - Fixed an IndexOutOfBoundsException with the entity limits feature